+++ apps/playground/next.config.mjs
@@ -1,6 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  transpilePackages: ['br-validators'],
+  transpilePackages: ['@br-validators/core'],
 };
 
 export default nextConfig;