+++ packages/br-validators/src/index.ts
@@ -1,6 +1,7 @@
 export * from './cnpj.js';
 export * from './cpf.js';
 export * from './cep.js';
+export * from './telefone.js';
 export * from './placa.js';
 export * from './pis-pasep.js';
 export {